// zoologico/src/app/models/employee.ts
export interface Employee {

  code?: number;
  nome?: string;
  sobrenome?: string;
  nuCpf?: string;
  
  enderecoCode?: number;
  pais?: string;
  estado?: string;
  cidade?: string;
  logradouro?: string;
  complemento?: string;
  
  zoologicoCode?: number;
  zoologicoCnpj?: string;
  zoologicoNome?: string;
  
  cargoCode?: number;
  cargoTitulo?: string;

  departamentoCode?: number;
  departamentoNome?: string;

}