// zoologico/src/app/models/zoo.ts
export interface Zoo {

  zoologicoCode?: number;
  zoologicoCnpj?: string;
  zoologicoNome?: string;
  funcionarioCode?: number;
  fornecedorCode?: number;
  fornecedorCnpj?: string;
  fornecedorRazaoSocial?: string;
  enderecoCode?: number;
  pais?: string;
  estado?: string;
  cidade?: string;
  logradouro?: string;
  complemento?: string;

}